requirements.txt
python-can>=4.2.0

